Refactor ReleaseChannel and GetChannelInfo into SetupLibrary, fix exception being thrown in recursive copy
