+++ src/AccessibilityInsights.SetupLibrary/ReleaseChannel.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace AccessibilityInsights.SharedUx.Enums
+namespace AccessibilityInsights.SetupLibrary
 {
     /// <summary>
     /// The release channel for a specific client. Channels appear in increasing upgrade frequency
