+++ src/AccessibilityInsights.SharedUx/ViewModels/ChannelConfigViewModel.cs
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
+using AccessibilityInsights.SetupLibrary;
 using AccessibilityInsights.SharedUx.Properties;
-using AccessibilityInsights.SharedUx.Enums;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
